Added a search filter
Added a search filter which is not case sensitive. Open Dashboard.js in components/dashboard/ to see the search query. We set the state of the query with a reducer in Navbar js and then get the state of that query in the dashboard so we can return the filtered projects without mutating the state or making another query to the database.